# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Livepeer Modules Transcode API
  version: 2.0.0
  description: Breaking paid-job/v1 and paid-session/v1 status surface.
paths:
...
  /api/v1/admin/waitlist/{id}/resend-verification:
    post:
      summary: Resend a pending unverified signup's verification email
      security: [{ AdminToken: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Provider accepted the email, or email is in dry-run mode; acceptance does not guarantee inbox delivery
          content:
            application/json:
              schema:
                type: object
                required: [status, delivery]
                properties:
                  status: { type: string, const: ok }
                  delivery: { type: string, enum: [accepted, dryrun] }
        "400": { description: Invalid signup ID }
        "401": { description: Admin authentication required }
        "409": { description: Signup is not pending and unverified }
        "429": { description: One resend per signup per minute per gateway instance }
        "502": { description: Provider failed; previous verification token remains valid }
...
  /api/v1/admin/operations:
    get:
      summary: List safe paid-operation diagnostics
      security: [{ AdminToken: [] }]
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [job, session] } }
        - { name: status, in: query, schema: { type: string, maxLength: 64 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 100 } }
      responses:
        "200":
          description: Most recently updated operations
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/AdminPaidOperationStatus" }
components:
  securitySchemes:
    UserSession: { type: http, scheme: bearer }
    ProductApiKey: { type: http, scheme: bearer }
    AdminToken: { type: http, scheme: bearer }
  schemas:
...
    PaidOperationStatus:
      type: object
      required: [operation_id, protocol, state, request_id, work_unit, funded_units, warnings, recovered, retry_count, output_state, output_status, output_state_since, last_failure_code]
      properties:
        operation_id: { type: string }
        protocol: { type: string, enum: [paid-job/v1, paid-session/v1] }
        state: { type: string }
        request_id: { type: string }
        work_unit: { type: string }
        funded_units: { type: string, pattern: "^(0|[1-9][0-9]*)$" }
        claimed_units: { type: [string, "null"], pattern: "^(0|[1-9][0-9]*)$" }
        balance_units: { type: [string, "null"], pattern: "^(0|[1-9][0-9]*)$" }
        lease_expires_at: { type: [string, "null"], format: date-time }
        warnings:
          type: array
          items: { type: string, enum: [refill_will_be_refused, funded_balance_exhausted, lease_expired, output_stalled] }
        recovered: { type: boolean }
        retry_count: { type: integer, minimum: 0 }
        next_retry_at: { type: [string, "null"], format: date-time }
        error_code: { type: [string, "null"] }
        relay_status: { type: [string, "null"] }
        delivered_units: { type: [string, "null"], pattern: "^(0|[1-9][0-9]*)$" }
        winddown_reason: { type: [string, "null"] }
        output_state: { type: [string, "null"], enum: [unknown, waiting, producing, stalled, null] }
        output_status: { type: [string, "null"], enum: [unknown, no_ingest, waiting, producing, stalled, output_failed, null] }
        output_state_since: { type: [string, "null"], format: date-time }
        last_failure_code: { type: [string, "null"], pattern: "^[a-z][a-z0-9_]{0,63}$" }
        terminal_at: { type: [string, "null"], format: date-time }
    AdminPaidOperationStatus:
      allOf:
        - { $ref: "#/components/schemas/PaidOperationStatus" }
        - type: object
          required: [kind, capability, offering, rotation_generation, settlement_sequence, created_at, updated_at]
          properties:
            kind: { type: string, enum: [job, session] }
            asset_id: { type: [string, "null"] }
            live_stream_id: { type: [string, "null"] }
            loc_operation_id: { type: [string, "null"] }
            broker_job_id: { type: [string, "null"] }
            broker_session_id: { type: [string, "null"] }
            capability: { type: string }
            offering: { type: string }
            transport: { type: [string, "null"], enum: [unary, stream, multipart, null] }
            quote_id: { type: string }
            quote_version: { type: string }
            route_fingerprint: { type: string }
            rotation_generation: { type: integer, minimum: 0 }
            settlement_sequence: { type: string, pattern: "^(0|[1-9][0-9]*)$" }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...